Rename dissassembler -> disassembler
