+++ DameDis/DisasmListCtrl.h
@@ -1,5 +1,5 @@
 #pragma once
-#include "Dissassembler.h"
+#include "Disassembler.h"
 #include <wx/listctrl.h>
 #include <vector>
